+++ test/Armada.Test.Unit/Program.cs
@@ -256,6 +256,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new RemoteTriggerEventHookTests());
             runner.AddSuite(new IncidentServiceTests());
             runner.AddSuite(new ObjectiveServiceTests());
+            runner.AddSuite(new KeyedAsyncLockTests());
             runner.AddSuite(new ObjectiveRefinementCoordinatorTests());
             runner.AddSuite(new ObjectiveDependencyAnalyzerTests());
             runner.AddSuite(new ObjectiveBriefRendererTests());